+++ __tests__/LottoAppTest.js
@@ -1,7 +1,6 @@
 import Lotto from "../src/Lotto.js";
 import LottoApp from "../src/LottoApp";
 import {
-  ERROR_MESSAGES,
   LOTTO_PRICE,
   LOTTO_NUMBERS,
   WINNING_PRIZES,
@@ -12,42 +11,6 @@ describe("LottoApp 클래스 테스트", () => {
 
   beforeEach(() => {
     lottoApp = new LottoApp();
-  });
-
-  test("구입 금액이 1,000원 단위가 아닌 경우 예외가 발생한다.", () => {
-    expect(() => {
-      lottoApp.validateAmount(1500);
-    }).toThrow(ERROR_MESSAGES.INVALID_PURCHASE_AMOUNT);
-  });
-
-  test("구입 금액이 0 이하인 경우 예외가 발생한다.", () => {
-    expect(() => {
-      lottoApp.validateAmount(0);
-    }).toThrow(ERROR_MESSAGES.INVALID_PURCHASE_AMOUNT);
-
-    expect(() => {
-      lottoApp.validateAmount(-1000);
-    }).toThrow(ERROR_MESSAGES.INVALID_PURCHASE_AMOUNT);
-  });
-
-  test("구입 금액이 숫자가 아닌 경우 예외가 발생한다.", () => {
-    expect(() => {
-      lottoApp.validateAmount("abc");
-    }).toThrow(ERROR_MESSAGES.INVALID_PURCHASE_AMOUNT);
-
-    expect(() => {
-      lottoApp.validateAmount(null);
-    }).toThrow(ERROR_MESSAGES.INVALID_PURCHASE_AMOUNT);
-  });
-
-  test("구입 금액이 유효한 경우 예외가 발생하지 않는다.", () => {
-    expect(() => {
-      lottoApp.validateAmount(1000);
-    }).not.toThrow();
-
-    expect(() => {
-      lottoApp.validateAmount(2000);
-    }).not.toThrow();
   });
 
   test("구입 금액에 따른 로또 개수가 정확히 계산된다.", () => {
@@ -64,36 +27,6 @@ describe("LottoApp 클래스 테스트", () => {
       expect(num).toBeGreaterThanOrEqual(LOTTO_NUMBERS.MIN);
       expect(num).toBeLessThanOrEqual(LOTTO_NUMBERS.MAX);
     });
-  });
-
-  test("보너스 번호가 숫자가 아닐 경우 예외가 발생한다.", () => {
-    expect(() => {
-      lottoApp.checkBonusNumberType("a");
-    }).toThrow(ERROR_MESSAGES.INVALID_BONUS_NUMBER_TYPE);
-  });
-
-  test("보너스 번호가 유효한 범위를 벗어날 경우 예외가 발생한다.", () => {
-    expect(() => {
-      lottoApp.checkBonusNumberRange(0);
-    }).toThrow(ERROR_MESSAGES.INVALID_BONUS_NUMBER_RANGE);
-
-    expect(() => {
-      lottoApp.checkBonusNumberRange(46);
-    }).toThrow(ERROR_MESSAGES.INVALID_BONUS_NUMBER_RANGE);
-  });
-
-  test("보너스 번호가 당첨 번호와 중복될 경우 예외가 발생한다.", () => {
-    const winningNumbers = [1, 2, 3, 4, 5, 6];
-    expect(() => {
-      lottoApp.checkBonusNumberDuplication(3, winningNumbers);
-    }).toThrow(ERROR_MESSAGES.DUPLICATE_BONUS_NUMBER);
-  });
-
-  test("보너스 번호가 1에서 45 사이의 유효한 번호이고 당첨 번호와 중복되지 않으면 예외가 발생하지 않는다.", () => {
-    const winningNumbers = [1, 2, 3, 4, 5, 6];
-    expect(() => {
-      lottoApp.validateBonusNumber(7, winningNumbers);
-    }).not.toThrow();
   });
 });
